% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pb_build}
\alias{pb_build}
\title{Build binary package}
\usage{
pb_build(ref, extra_dependencies = NULL, workdir = NULL)
}
\arguments{
\item{ref}{A remote-style refrence to a package (e.g., \code{user/repo})}

\item{extra_dependencies}{An optional character vector of
additional dependencies to install, also remote-style references.}

\item{workdir}{An optional working directory to use. If non-NULL,
then source trees will persist across calls to \code{pb_build}, which
will make things a little faster.}
}
\description{
Build a binary package, given a remote-style reference to a
package (e.g., \code{user/repo})
}
