% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pb_server}
\alias{pb_server}
\title{Run a pkgbuilder server}
\usage{
pb_server(versions, workdir, port)
}
\arguments{
\item{versions}{A vector of 2-digit version numbers corresponding
to the versions of R to build packages for.  Workers will need
to be provided separately to build the packages.}

\item{workdir}{A directory to work in. This must be shared between
the workers and the server, so do not use a temporary directory.}

\item{port}{The port to run on. Your operating system will likely
restrict you from using very low port numbers, so use something
like 8080}
}
\value{
Never returns - runs a HTTP server as a side-effect
}
\description{
Run a pkgbuilder server
}
