% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pb_source}
\alias{pb_source}
\title{Update package source}
\usage{
pb_source(ref, workdir = NULL)
}
\arguments{
\item{ref}{A remote-style refrence to a package (e.g., \code{user/repo})}

\item{workdir}{An optional working directory to use. If non-NULL,
then source trees will persist across calls to \code{pb_build}, which
will make things a little faster.}
}
\value{
Character string, being the path to the checked out source tree
}
\description{
Update a package source
}
