% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{pb_worker}
\alias{pb_worker}
\title{Run a pkgbuilder worker}
\usage{
pb_worker(workdir, timeout = NULL)
}
\arguments{
\item{workdir}{Working directory for the queue - must be shared
with the server process.}

\item{timeout}{A per-job timeout.  Builds that take longer than
this many seconds will be terminated.}
}
\value{
Never returns, runs a worker as a side-effect
}
\description{
Run a pkgbuilder worker
}
